refactor: :truck: Change `Transfer` page and story to `Send`
