--- src/ui/pages/Transfer/Transfer.stories.tsx
@@ -1,9 +1,0 @@
-import { ComponentMeta, ComponentStory } from '@storybook/react'
-import { Transfer } from './index'
-
-export default {
-  title: 'Pages/Transfer',
-  component: Transfer
-} as ComponentMeta<typeof Transfer>
-
-export const Playground: ComponentStory<typeof Transfer> = () => <Transfer />
